requirements.txt

google-cloud-bigquery
